Scalajs 1.0.0 support (#1157)
* ScalaJS 1.0.0 support

* fix intercept errors

* fix intercept errors

* Check both JS versions in CI

* fixed missed intercept

* fixed missed intercept

* Remove SCALAJS_VERSION from sbt command in travis

* Use ScalaJS 0.6.32 for Scala 2.11 in CI

* Move UncaughtExceptionHandler test to JVM only

* fix hardcoded scalajs version in plugins.sbt

* Update Cats-Effect version

+++ monix-execution/js/src/main/scala/monix/execution/schedulers/JSTimer.scala
@@ -30,7 +30,7 @@ private[schedulers] object JSTimer {
         r.run()
       } catch { case t: Throwable => ec.reportFailure(t) }
 
-    js.Dynamic.global.setTimeout(lambda, delayMillis)
+    js.Dynamic.global.setTimeout(lambda, delayMillis.toDouble)
   }
 
   def clearTimeout(task: js.Dynamic): js.Dynamic = {
